src/main/java/com/outgrowthsolutions/ogsrecipeapp/converters/UnitOfMeasureToUnitOfMeasureCommand.java
package com.outgrowthsolutions.ogsrecipeapp.converters;

import com.outgrowthsolutions.ogsrecipeapp.commands.UnitOfMeasureCommand;
import com.outgrowthsolutions.ogsrecipeapp.domain.UnitOfMeasure;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class UnitOfMeasureToUnitOfMeasureCommand implements Converter<UnitOfMeasure, UnitOfMeasureCommand> {
    @Override
    public UnitOfMeasureCommand convert(UnitOfMeasure unitOfMeasure) {

        if (unitOfMeasure == null)
            return null;

        final UnitOfMeasureCommand converted = new UnitOfMeasureCommand();
        converted.setId(unitOfMeasure.getId());
        converted.setDescription(unitOfMeasure.getDescription());
        return converted;
    }
}
